+++ test/unittest/component/test_table.py
@@ -1,0 +1,7 @@
+from superduper import Schema, Table
+
+
+def test_calls_post_init():
+    t = Table('test', schema=Schema('test', fields={'x': 'str'}))
+
+    assert hasattr(t, 'version')